+++ packages/nimble-components/src/rich-text/viewer/tests/rich-text-viewer.spec.ts
@@ -2,11 +2,55 @@ import { html } from '@microsoft/fast-element';
 import { RichTextViewer, richTextViewerTag } from '..';
 import { fixture, type Fixture } from '../../../utilities/tests/fixture';
 import { RichTextViewerPageObject } from '../testing/rich-text-viewer.pageobject';
+import { anchorTag } from '../../../anchor';
+import { waitForUpdatesAsync } from '../../../testing/async-helpers';
+import {
+    RichTextMentionUsers,
+    richTextMentionUsersTag
+} from '../../../rich-text-mention/users';
+import { richTextMentionUsersViewTag } from '../../../rich-text-mention/users/view';
+import { MappingUser, mappingUserTag } from '../../../mapping/user';
 
 async function setup(): Promise<Fixture<RichTextViewer>> {
     return fixture<RichTextViewer>(
-        html`<nimble-rich-text-viewer></nimble-rich-text-viewer>`
+        html`<${richTextViewerTag}></${richTextViewerTag}>`
     );
+}
+
+async function setupMentionConfig(): Promise<Fixture<RichTextViewer>> {
+    return fixture<RichTextViewer>(
+        // prettier-ignore
+        html`<${richTextViewerTag}>
+            <${richTextMentionUsersTag} pattern="^user:(.*)">
+                <${mappingUserTag} key="user:1" display-name="John Doe"></${mappingUserTag}>
+                <${mappingUserTag} key="user:2" display-name="Mary Wilson"></${mappingUserTag}>
+            </${richTextMentionUsersTag}>
+        </${richTextViewerTag}>`
+    );
+}
+
+async function appendUserMentionConfiguration(
+    element: RichTextViewer,
+    userKeys?: string[],
+    displayNames?: string[]
+): Promise<void> {
+    const userMention = document.createElement(
+        richTextMentionUsersTag
+    ) as RichTextMentionUsers;
+    userMention.pattern = '^user:(.*)';
+
+    if (userKeys || displayNames) {
+        userKeys?.forEach((userKey, index) => {
+            const mappingUser = document.createElement(
+                mappingUserTag
+            ) as MappingUser;
+            mappingUser.key = userKey ?? '';
+            mappingUser.displayName = displayNames?.[index] ?? '';
+            userMention.appendChild(mappingUser);
+        });
+    }
+    element.appendChild(userMention);
+    await waitForUpdatesAsync();
 }
 
 describe('RichTextViewer', () => {
@@ -21,9 +65,9 @@ describe('RichTextViewer', () => {
     });
 
     it('can construct an element instance', () => {
-        expect(
-            document.createElement('nimble-rich-text-viewer')
-        ).toBeInstanceOf(RichTextViewer);
+        expect(document.createElement(richTextViewerTag)).toBeInstanceOf(
+            RichTextViewer
+        );
     });
 
     it('should export its tag', () => {
@@ -80,5 +124,299 @@ describe('RichTextViewer', () => {
         );
 
         await disconnect();
+    });
+
+    describe('user mention dynamic loading', () => {
+        beforeEach(async () => {
+            await connect();
+        });
+
+        afterEach(async () => {
+            await disconnect();
+        });
+
+        it('adding mention configuration converts the absolute link matching the pattern to mention node', async () => {
+            element.markdown = '<user:1>';
+
+            expect(pageObject.getRenderedMarkdownTagNames()).toEqual([
+                'P',
+                `${anchorTag}`.toUpperCase()
+            ]);
+            expect(pageObject.getRenderedMarkdownLastChildContents()).toBe(
+                'user:1'
+            );
+            await appendUserMentionConfiguration(element);
+
+            expect(pageObject.getRenderedMarkdownTagNames()).toEqual([
+                'P',
+                `${richTextMentionUsersViewTag}`.toUpperCase()
+            ]);
+            expect(
+                pageObject.getRenderedMarkdownLastChildAttribute(
+                    'mention-label'
+                )
+            ).toEqual('1');
+        });
+
+        // TODO: Once the rich text validator added for duplicate configuration elements, below test case should be updated
+        it('adding two mention configuration elements in the same viewer should render as mention view', async () => {
+            element.markdown = '<user:1>';
+            await appendUserMentionConfiguration(element);
+            await appendUserMentionConfiguration(element);
+
+            expect(pageObject.getRenderedMarkdownTagNames()).toEqual([
+                'P',
+                `${richTextMentionUsersViewTag}`.toUpperCase()
+            ]);
+            expect(
+                pageObject.getRenderedMarkdownLastChildAttribute(
+                    'mention-label'
+                )
+            ).toEqual('1');
+        });
+
+        it('adding mention mapping elements renders the mapped display name', async () => {
+            element.markdown = '<user:1>';
+            await appendUserMentionConfiguration(
+                element,
+                ['user:1'],
+                ['username1']
+            );
+
+            expect(pageObject.getRenderedMarkdownTagNames()).toEqual([
+                'P',
+                `${richTextMentionUsersViewTag}`.toUpperCase()
+            ]);
+            expect(
+                pageObject.getRenderedMarkdownLastChildAttribute(
+                    'mention-label'
+                )
+            ).toEqual('username1');
+        });
+
+        it('adding two mention mapping elements renders the mapped display names', async () => {
+            element.markdown = '<user:1> <user:2>';
+            await appendUserMentionConfiguration(
+                element,
+                ['user:1', 'user:2'],
+                ['username1', 'username2']
+            );
+
+            expect(pageObject.getRenderedMarkdownTagNames()).toEqual([
+                'P',
+                `${richTextMentionUsersViewTag}`.toUpperCase(),
+                `${richTextMentionUsersViewTag}`.toUpperCase()
+            ]);
+            expect(
+                pageObject.getRenderedMarkdownAttributeValues('mention-label')
+            ).toEqual(['username1', 'username2']);
+        });
+
+        it('removing configuration element renders the mention node as absolute link', async () => {
+            element.markdown = '<user:1>';
+            await appendUserMentionConfiguration(
+                element,
+                ['user:1'],
+                ['username1']
+            );
+
+            const renderedUserMention = element.firstElementChild as RichTextMentionUsers;
+            element.removeChild(renderedUserMention);
+            await waitForUpdatesAsync();
+
+            expect(pageObject.getRenderedMarkdownTagNames()).toEqual([
+                'P',
+                `${anchorTag}`.toUpperCase()
+            ]);
+            expect(pageObject.getRenderedMarkdownLastChildContents()).toBe(
+                'user:1'
+            );
+        });
+
+        it('removing mapping element renders the mention node with user ID', async () => {
+            element.markdown = '<user:1>';
+            await appendUserMentionConfiguration(
+                element,
+                ['user:1'],
+                ['username1']
+            );
+
+            const renderedUserMention = element.firstElementChild as RichTextMentionUsers;
+            const renderedMappingUser = renderedUserMention.firstElementChild as MappingUser;
+            renderedUserMention.removeChild(renderedMappingUser);
+            await waitForUpdatesAsync();
+
+            expect(pageObject.getRenderedMarkdownTagNames()).toEqual([
+                'P',
+                `${richTextMentionUsersViewTag}`.toUpperCase()
+            ]);
+            expect(
+                pageObject.getRenderedMarkdownLastChildAttribute(
+                    'mention-label'
+                )
+            ).toEqual('1');
+        });
+
+        it('updating to the `pattern` in mention configuration converts the mention to absolute link', async () => {
+            element.markdown = '<user:1>';
+            await appendUserMentionConfiguration(
+                element,
+                ['user:1'],
+                ['username1']
+            );
+
+            (element.firstElementChild as RichTextMentionUsers).pattern = 'invalid';
+            await waitForUpdatesAsync();
+
+            expect(pageObject.getRenderedMarkdownTagNames()).toEqual([
+                'P',
+                `${anchorTag}`.toUpperCase()
+            ]);
+            expect(pageObject.getRenderedMarkdownLastChildContents()).toBe(
+                'user:1'
+            );
+        });
+
+        it('updating `display-name` in mapping mention should update the `mention-label` in view', async () => {
+            element.markdown = '<user:1>';
+            await appendUserMentionConfiguration(
+                element,
+                ['user:1'],
+                ['username1']
+            );
+
+            const renderedUserMention = element.firstElementChild as RichTextMentionUsers;
+            const renderedMappingUser = renderedUserMention.firstElementChild as MappingUser;
+            renderedMappingUser.displayName = 'updated-name';
+            await waitForUpdatesAsync();
+
+            expect(pageObject.getRenderedMarkdownTagNames()).toEqual([
+                'P',
+                `${richTextMentionUsersViewTag}`.toUpperCase()
+            ]);
+            expect(
+                pageObject.getRenderedMarkdownLastChildAttribute(
+                    'mention-label'
+                )
+            ).toEqual('updated-name');
+        });
+
+        it('updating valid `key` in mapping mention should update it to a mention view if it is a absolute link before', async () => {
+            element.markdown = '<user:2>';
+            await appendUserMentionConfiguration(
+                element,
+                ['invalid'],
+                ['username']
+            );
+
+            expect(pageObject.getRenderedMarkdownTagNames()).toEqual([
+                'P',
+                `${anchorTag}`.toUpperCase()
+            ]);
+            expect(pageObject.getRenderedMarkdownLastChildContents()).toBe(
+                'user:2'
+            );
+
+            const renderedUserMention = element.firstElementChild as RichTextMentionUsers;
+            const renderedMappingUser = renderedUserMention.firstElementChild as MappingUser;
+            renderedMappingUser.key = 'user:2';
+            await waitForUpdatesAsync();
+
+            expect(pageObject.getRenderedMarkdownTagNames()).toEqual([
+                'P',
+                `${richTextMentionUsersViewTag}`.toUpperCase()
+            ]);
+            expect(
+                pageObject.getRenderedMarkdownLastChildAttribute(
+                    'mention-label'
+                )
+            ).toEqual('username');
+        });
+
+        it('updating invalid `key` in mapping mention should update it to absolute link if it is a mention before', async () => {
+            element.markdown = '<user:2>';
+            await appendUserMentionConfiguration(
+                element,
+                ['user:2'],
+                ['username']
+            );
+
+            expect(pageObject.getRenderedMarkdownTagNames()).toEqual([
+                'P',
+                `${richTextMentionUsersViewTag}`.toUpperCase()
+            ]);
+            expect(
+                pageObject.getRenderedMarkdownLastChildAttribute(
+                    'mention-label'
+                )
+            ).toEqual('username');
+
+            const renderedUserMention = element.firstElementChild as RichTextMentionUsers;
+            const renderedMappingUser = renderedUserMention.firstElementChild as MappingUser;
+            renderedMappingUser.key = 'invalid';
+            await waitForUpdatesAsync();
+
+            expect(pageObject.getRenderedMarkdownTagNames()).toEqual([
+                'P',
+                `${anchorTag}`.toUpperCase()
+            ]);
+            expect(pageObject.getRenderedMarkdownLastChildContents()).toBe(
+                'user:2'
+            );
+        });
+    });
+
+    describe('user mention via template', () => {
+        beforeEach(async () => {
+            ({ element, connect, disconnect } = await setupMentionConfig());
+            pageObject = new RichTextViewerPageObject(element);
+            await connect();
+        });
+
+        afterEach(async () => {
+            await disconnect();
+        });
+
+        it('should render as mention view element when the input markdown matching the mention string with pattern', async () => {
+            element.markdown = '<user:1>';
+            await waitForUpdatesAsync();
+
+            expect(pageObject.getRenderedMarkdownTagNames()).toEqual([
+                'P',
+                `${richTextMentionUsersViewTag}`.toUpperCase()
+            ]);
+            expect(
+                pageObject.getRenderedMarkdownLastChildAttribute(
+                    'mention-label'
+                )
+            ).toEqual('John Doe');
+        });
+
+        it('should render as absolute link when the input markdown does not match with pattern', async () => {
+            element.markdown = '<https://user/1>';
+            await waitForUpdatesAsync();
+
+            expect(pageObject.getRenderedMarkdownTagNames()).toEqual([
+                'P',
+                `${anchorTag}`.toUpperCase()
+            ]);
+            expect(pageObject.getRenderedMarkdownLastChildContents()).toBe(
+                'https://user/1'
+            );
+        });
+
+        it('should render as mention view elements when multiple mention strings are passed in the markdown string', async () => {
+            element.markdown = '<user:1> <user:2>';
+            await waitForUpdatesAsync();
+
+            expect(pageObject.getRenderedMarkdownTagNames()).toEqual([
+                'P',
+                `${richTextMentionUsersViewTag}`.toUpperCase(),
+                `${richTextMentionUsersViewTag}`.toUpperCase()
+            ]);
+            expect(
+                pageObject.getRenderedMarkdownAttributeValues('mention-label')
+            ).toEqual(['John Doe', 'Mary Wilson']);
+        });
     });
 });